feat: add hamburger menu
