+++ src/components/VideoCardPanel/VideoCardPanel.component.jsx
@@ -3,9 +3,12 @@ import VideoCard from '../VideoCard';
 import './VideoCardPanel.styles.css';
 
 function VideoCardPanel({ videos }) {
-  const getVideoThumbnail = (video) => video.snippet ? video.snippet.thumbnails.medium.url : video.videoThumbnail;
-  const getVideoTitle = (video) => video.snippet ? video.snippet.title: video.videoTitle;
-  const getVideoDescription = (video) => video.snippet ? video.snippet.description: video.videoDescription;
+  const getVideoThumbnail = (video) =>
+    video.snippet ? video.snippet.thumbnails.medium.url : video.videoThumbnail;
+  const getVideoTitle = (video) =>
+    video.snippet ? video.snippet.title : video.videoTitle;
+  const getVideoDescription = (video) =>
+    video.snippet ? video.snippet.description : video.videoDescription;
 
   return (
     <div className="video-card-wrapper">
